refactor: decompose frontend components and fix code quality issues
- Split large components into single-responsibility units:
  StepReviewIssue (1101→12 lines), StepCreateCredentials (418→130 lines),
  EventForm (724→290 lines), dashboard page, header, file-dropzone, events page

- Extract reusable hooks: useEventTypes, useEventLevels, useEventFormats,
  useEventFormInitialization, useMultiSelectFilters, useRecipientValidation

- Extract sub-components: CredentialLayout, CredentialTable, CredentialCards,
  CredentialStatusBadge, CredentialDetailView, IssueCredentialView,
  preview/*, EventFormStep0, EventFormStep1, FilterTagList,
  StatsCard, RecentEventsList, HeaderAvatar, dropzone/*

- Fix BUG-021: replace all hardcoded route strings with ROUTES.* / createRoute.*
- Fix BUG-022: remove duplicate error toasts (catch blocks + global MutationCache)
- Fix BUG-023: remove duplicate formatDate — import from CredentialLayout
- Fix BUG-024: use showErrorToast (not showSuccessToast) for validation errors
- Fix BUG-025: use useUser() selector instead of full useAuthStore() subscription

- Add missing ROUTES.SETTINGS.ARCHIVED and createRoute.publicRecipientDetail
- Update FRONTEND_BUGS.md with BUG-021 to BUG-033 (13 new entries, 5 fixed)
