+++ apps/frontend/src/app/(dashboard)/credentials/page.tsx
@@ -7,42 +7,10 @@ import { useRouter, useSearchParams } from 'next/navigation';
 import { useCredentials } from '@/hooks/useCredentials';
 import { useEvents } from '@/hooks/useEvents';
 import { Pagination, ModulePermissionError } from '@/components';
-import { ROUTES, createRoute } from '@/config/routes';
-import { CredentialStatus } from '@/types/credential.types';
+import { ROUTES } from '@/config/routes';
 import { ListFilter } from 'lucide-react';
-
-const STATUS_CONFIG: Record<
-  CredentialStatus,
-  { label: string; dotColor: string; textColor: string }
-> = {
-  [CredentialStatus.ISSUED]: {
-    label: 'Issued',
-    dotColor: 'bg-green-500',
-    textColor: 'text-green-700',
-  },
-  [CredentialStatus.PENDING]: {
-    label: 'Pending',
-    dotColor: 'bg-yellow-400',
-    textColor: 'text-yellow-700',
-  },
-  [CredentialStatus.PROCESSING]: {
-    label: 'Processing',
-    dotColor: 'bg-blue-500',
-    textColor: 'text-blue-700',
-  },
-  [CredentialStatus.FAILED]: { label: 'Failed', dotColor: 'bg-red-500', textColor: 'text-red-700' },
-  [CredentialStatus.DRAFT]: { label: 'Draft', dotColor: 'bg-gray-400', textColor: 'text-gray-500' },
-};
-
-function CredentialStatusBadge({ status }: { status: CredentialStatus }) {
-  const config = STATUS_CONFIG[status];
-  return (
-    <span className={`inline-flex items-center gap-1.5 text-sm ${config.textColor}`}>
-      <span className={`w-2 h-2 rounded-full ${config.dotColor}`} />
-      {config.label}
-    </span>
-  );
-}
+import { CredentialTable } from '@/components/credentials/CredentialTable';
+import { CredentialCards } from '@/components/credentials/CredentialCards';
 
 const ITEMS_PER_PAGE = 10;
 
@@ -109,15 +77,6 @@ export default function CredentialsPage() {
   const handleEventFilter = (eventId: string) => {
     updateParams({ eventId, page: '' });
     setIsFilterOpen(false);
-  };
-
-  const formatDate = (date: string | null) => {
-    if (!date) return '--';
-    return new Date(date).toLocaleDateString('en-GB', {
-      day: 'numeric',
-      month: 'long',
-      year: 'numeric',
-    });
   };
 
   if (error?.message.includes('403')) {
@@ -197,131 +156,8 @@ export default function CredentialsPage() {
           </div>
         </div>
 
-        {/* Desktop Table */}
-        <div className="hidden sm:block overflow-x-auto">
-          <table className="w-full text-sm text-left text-gray-500">
-            <thead className="text-left font-black text-gray-600 bg-gray-50 border-b border-gray-200">
-              <tr>
-                <th scope="col" className="px-6 py-3 font-medium">
-                  Name
-                </th>
-                <th scope="col" className="px-6 py-3 font-medium">
-                  Email
-                </th>
-                <th scope="col" className="px-6 py-3 font-medium">
-                  Events
-                </th>
-                <th scope="col" className="px-6 py-3 font-medium">
-                  Issue Date
-                </th>
-                <th scope="col" className="px-6 py-3 font-medium">
-                  Status
-                </th>
-                <th scope="col" className="px-6 py-3 font-medium">
-                  Action
-                </th>
-              </tr>
-            </thead>
-            <tbody>
-              {isLoading ? (
-                <tr>
-                  <td colSpan={6} className="px-6 py-12 text-center">
-                    <div className="flex justify-center">
-                      <div className="animate-spin rounded-full h-6 w-6 border-b-2 border-blue-600" />
-                    </div>
-                  </td>
-                </tr>
-              ) : credentials.length === 0 ? (
-                <tr>
-                  <td colSpan={6} className="px-6 py-12 text-center text-gray-500">
-                    No credentials found
-                  </td>
-                </tr>
-              ) : (
-                credentials.map((item) => (
-                  <tr key={item.uuid} className="border-b border-gray-200 hover:bg-gray-50">
-                    <td className="px-6 py-2 text-gray-900">{item.recipient?.name}</td>
-                    <td className="px-6 py-3 text-gray-500">{item.recipient?.email}</td>
-                    <td className="px-6 py-3">
-                      {item.event ? (
-                        <button
-                          onClick={() => router.push(createRoute.eventDetail(item.event!.uuid))}
-                          className="text-gray-900 hover:text-blue-600 hover:underline cursor-pointer"
-                        >
-                          {item.event.name}
-                        </button>
-                      ) : (
-                        <span className="text-gray-400">--</span>
-                      )}
-                    </td>
-                    <td className="px-6 py-2 text-gray-900">{formatDate(item.issued_date)}</td>
-                    <td className="px-6 py-2">
-                      <CredentialStatusBadge status={item.status} />
-                    </td>
-                    <td className="px-6 py-3">
-                      <a
-                        href={createRoute.credentialDetail(item.uuid)}
-                        onClick={(e) => {
-                          e.preventDefault();
-                          router.push(createRoute.credentialDetail(item.uuid));
-                        }}
-                        className="text-blue-600 hover:underline font-medium"
-                      >
-                        View
-                      </a>
-                    </td>
-                  </tr>
-                ))
-              )}
-            </tbody>
-          </table>
-        </div>
-
-        {/* Mobile Cards */}
-        <div className="sm:hidden divide-y divide-gray-200">
-          {isLoading ? (
-            <div className="flex justify-center py-12">
-              <div className="animate-spin rounded-full h-6 w-6 border-b-2 border-blue-600" />
-            </div>
-          ) : credentials.length === 0 ? (
-            <p className="text-center py-12 text-gray-500 text-sm">No credentials found</p>
-          ) : (
-            credentials.map((item) => (
-              <div key={item.uuid} className="block px-4 py-2 hover:bg-gray-50">
-                <div className="flex items-center justify-between mb-1.5">
-                  <span className="font-semibold text-sm text-gray-900">
-                    {item.recipient?.name}
-                  </span>
-                  <CredentialStatusBadge status={item.status} />
-                </div>
-                <p className="text-xs text-gray-500 mb-2">{item.recipient?.email}</p>
-                <div className="flex items-center gap-4 text-xs text-gray-500">
-                  <span>
-                    {item.event ? (
-                      <button
-                        onClick={() => router.push(createRoute.eventDetail(item.event!.uuid))}
-                        className="text-blue-600 hover:underline cursor-pointer"
-                      >
-                        {item.event.name}
-                      </button>
-                    ) : (
-                      '--'
-                    )}
-                  </span>
-                  <span>{formatDate(item.issued_date)}</span>
-                  <a
-                    href={createRoute.publicCredential(item.uuid)}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    className="ml-auto text-blue-600 hover:underline font-medium"
-                  >
-                    View
-                  </a>
-                </div>
-              </div>
-            ))
-          )}
-        </div>
+        <CredentialTable credentials={credentials} isLoading={isLoading} />
+        <CredentialCards credentials={credentials} isLoading={isLoading} />
       </div>
       {/* Pagination  */}
       {meta && meta.totalPages > 1 && (
