+++ apps/frontend/src/app/(dashboard)/dashboard/page.tsx
@@ -1,92 +1,14 @@
 'use client';
 
-import { useAuthStore } from '@/store/auth.store';
+import { useUser } from '@/store/auth.store';
 import { useEvents } from '@/hooks/useEvents';
 import { useCredentials } from '@/hooks/useCredentials';
 import { useDesignList } from '@/hooks/useDesigns';
-import { ROUTES, createRoute } from '@/config/routes';
-import type { Event } from '@/types';
-import { Calendar, Users, Award, Palette, Loader2 } from 'lucide-react';
+import { ROUTES } from '@/config/routes';
+import { Calendar, Users, Award, Palette } from 'lucide-react';
 import Link from 'next/link';
-
-/**
- * Dashboard Stats Card
- */
-function StatsCard({
-  title,
-  value,
-  icon: Icon,
-  isLoading,
-}: {
-  title: string;
-  value: string | number;
-  icon: React.ComponentType<{ className?: string }>;
-  isLoading?: boolean;
-}) {
-  return (
-    <div className="bg-white rounded-lg shadow-sm border border-gray-200 p-6">
-      <div className="flex items-center justify-between">
-        <div>
-          <p className="text-sm font-medium text-gray-500">{title}</p>
-          {isLoading ? (
-            <Loader2 className="mt-2 w-6 h-6 animate-spin text-gray-400" />
-          ) : (
-            <p className="mt-1 text-3xl font-semibold text-gray-900">{value}</p>
-          )}
-        </div>
-        <div className="p-3 bg-primary-100 rounded-lg">
-          <Icon className="w-6 h-6 text-primary-600" />
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Recent Events List
- */
-function RecentEventsList({ events, isLoading }: { events: Event[]; isLoading: boolean }) {
-  if (isLoading) {
-    return (
-      <div className="flex justify-center py-8">
-        <Loader2 className="w-6 h-6 animate-spin text-gray-400" />
-      </div>
-    );
-  }
-
-  if (events.length === 0) {
-    return (
-      <p className="text-center text-gray-500 py-8">
-        No events yet. Create your first event to get started.
-      </p>
-    );
-  }
-
-  return (
-    <div className="space-y-4">
-      {events.slice(0, 5).map((event) => (
-        <div
-          key={event.uuid}
-          className="flex items-center justify-between p-4 bg-gray-50 rounded-lg"
-        >
-          <div className="min-w-0 flex-1">
-            <p className="font-medium text-gray-900 truncate">{event.name}</p>
-            <p className="text-sm text-gray-500">
-              {event.event_type?.name ?? 'No type'} &bull;{' '}
-              {new Date(event.createdAt).toLocaleDateString()}
-            </p>
-          </div>
-          <Link
-            href={createRoute.eventDetail(event.uuid)}
-            className="ml-4 text-sm text-primary-600 hover:text-primary-700 font-medium shrink-0"
-          >
-            View
-          </Link>
-        </div>
-      ))}
-    </div>
-  );
-}
+import { StatsCard } from '@/components/dashboard/StatsCard';
+import { RecentEventsList } from '@/components/dashboard/RecentEventsList';
 
 /**
  * Dashboard Page
@@ -94,7 +16,7 @@ function RecentEventsList({ events, isLoading }: { events: Event[]; isLoading: b
  * Main dashboard for authenticated users showing real data from APIs
  */
 export default function DashboardPage() {
-  const { user } = useAuthStore();
+  const user = useUser();
 
   const { data: eventsData, isLoading: eventsLoading } = useEvents({
     limit: 5,
