+++ apps/frontend/src/components/ui/file-dropzone.tsx
@@ -4,6 +4,9 @@ import { useState, useRef, useCallback, useEffect } from 'react';
 import { FileCategory } from '@/services/api/file.service';
 import { deleteImage, replaceImage } from '@/lib/image-upload';
 import { getApiErrorMessage } from '@/lib/api-error';
+import { DropzonePlaceholder } from './dropzone/DropzonePlaceholder';
+import { DropzoneImageError } from './dropzone/DropzoneImageError';
+import { UploadSpinner } from './dropzone/UploadSpinner';
 
 /**
  * File Dropzone Component Props
@@ -300,32 +303,7 @@ export function FileDropzone({
             />
 
             {/* Error state */}
-            {imageError && !imageLoading && (
-              <div
-                className={`absolute inset-0 flex items-center justify-center bg-gray-100 rounded-md border border-gray-200 ${
-                  imageSize === 'large' ? 'min-w-[200px] min-h-[120px]' : ''
-                }`}
-              >
-                <div className="text-center p-2">
-                  <svg
-                    className={`${
-                      imageSize === 'small' ? 'w-6 h-6' : 'w-8 h-8'
-                    } text-gray-400 mx-auto mb-1`}
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-                    />
-                  </svg>
-                  <p className="text-xs text-gray-500">Failed to load</p>
-                </div>
-              </div>
-            )}
+            {imageError && !imageLoading && <DropzoneImageError imageSize={imageSize} />}
 
             {/* Remove button - only show when image is loaded */}
             {!imageLoading && !imageError && (
@@ -377,52 +355,9 @@ export function FileDropzone({
                             }`}
           >
             {uploading ? (
-              <>
-                <svg
-                  className="w-10 h-10 mb-3 text-gray-400 animate-spin"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                >
-                  <circle
-                    className="opacity-25"
-                    cx="12"
-                    cy="12"
-                    r="10"
-                    stroke="currentColor"
-                    strokeWidth="4"
-                  />
-                  <path
-                    className="opacity-75"
-                    fill="currentColor"
-                    d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-                  />
-                </svg>
-                <p className="text-sm text-gray-500">Uploading...</p>
-              </>
+              <UploadSpinner />
             ) : (
-              <>
-                {/* Upload Icon */}
-                <svg
-                  className="w-10 h-10 mb-2 text-gray-400"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={1.5}
-                    d="M7 16a4 4 0 01-.88-7.903A5 5 0 1115.9 6L16 6a5 5 0 011 9.9M15 13l-3-3m0 0l-3 3m3-3v12"
-                  />
-                </svg>
-                <p className="mb-2 text-sm text-gray-500">
-                  <span className="font-semibold text-black-600">Click to upload</span> or drag and
-                  drop
-                </p>
-                <p className="text-xs text-gray-500">
-                  {category === 'banner' ? 'Size: 1920x300' : `Max. File Size: ${maxSizeMB}MB`}
-                </p>
-              </>
+              <DropzonePlaceholder category={category} maxSizeMB={maxSizeMB} />
             )}
           </div>
         </div>
